Restructure code
Used PSA implementation as an orientation
